fix(lib/babe): check if authority index is in the `authorities` range (#2601)

+++ lib/babe/verify.go
@@ -390,7 +390,7 @@ func (b *verifier) verifyPreRuntimeDigest(digest *types.PreRuntimeDigest) (scale
 		authIdx = d.AuthorityIndex
 	}
 
-	if len(b.authorities) <= int(authIdx) {
+	if uint64(len(b.authorities)) <= uint64(authIdx) {
 		logger.Tracef("verifyPreRuntimeDigest invalid auth index %d, we have %d auths",
 			authIdx, len(b.authorities))
 		return nil, ErrInvalidBlockProducerIndex
